Set up bare-bones Angular app and Node backend
Angular app set up with minimal login and feed pages. Set up a Node.js
server in /backend to handle API requests and serve placeholder data for
now.  Currently, the app will show a login page and feed page at '#/'
and '#/feed' respectively. On the feed page is a button to send a GET
request to the node backend and display the list of developers it
returns.