Implemented voting backend, some front-end for testing
Works from poll page. Does not yet from feed (should be simple to write
feed controller function with index).

When a vote is submitted with the user’s netid,
-Checks db to see if user has voted on poll already
-If has, then same vote negates it. Different vote changes their vote.
-Otherwise, add new vote object to Vote collection
-Then updates upvote and down vote counts, updates score

Need to do: Modularize function (it’s bloated and gigantic), make sure
that the async part is right (non-guaranteed behavior atm)