Implemented get10, sorting
Sorting for both Best and Newest works!

feedController makes a call to neither 10Best or 10New, and can specify
a “num” which is the number of polls which have been grabbed and
displayed on the page already. The polls gotten will start from there.

Currently feed just shows the polls in order. Code for votes and
responses are commented out. Wouldn’t display properly with them.
Hopefully can be fixed relatively easily.

Sidetone: Using async package. Installed it on DigitalOcean server
already.